// workspaces/mooncg/src/types/mooncg.ts
type Person =
	| {
			name: string;
			email?: string;
			url?: string;
	  }
	| string;

export namespace MoonCG {
	/**
	 * A collection of types that represent the raw data from the `mooncg` stanza in a bundle's `package.json`.
	 */
	export namespace Manifest {
		/**
		 * An unparsed `assetCategory` from a bundle's `package.json`.
		 */
		export interface UnparsedAssetCategory {
			name: string;
			title: string;
			allowedTypes?: string[];
		}

		/**
		 * An unparsed `panel` from a bundle's `package.json`.
		 */
		export interface UnparsedPanel {
			name: string;
			title: string;
			file: string;
			headerColor?: string;
			fullbleed?: boolean;
			workspace?: string;
			dialog?: boolean;
			dialogButtons?: { name: string; type: "dismiss" | "confirm" }[];
			width?: number;
		}

		/**
		 * An unparsed `graphic` from a bundle's `package.json`.
		 */
		export interface UnparsedGraphic {
			file: string;
			width: number;
			height: number;
			singleInstance?: boolean;
		}

		/**
		 * An unparsed `mount` configuration from a bundle's `package.json`.
		 */
		export interface UnparsedMount {
			directory: string;
			endpoint: string;
		}

		/**
		 * An unparsed `soundCue` from a bundle's `package.json`.
		 */
		export interface UnparsedSoundCue {
			name: string;
			assignable?: boolean;
			defaultVolume?: number;
			defaultFile?: string;
		}

		/**
		 * An unparsed record of `bundleDependencies` from a bundle's `package.json`.
		 */
		export type UnparsedBundleDependencies = Record<string, string>;

		/**
		 * The actual structure of a bundle's `mooncg` stanza from it's `package.json`.
		 */
		export interface UnparsedManifest {
			compatibleRange?: string;
			transformBareModuleSpecifiers?: boolean;
			dashboardPanels?: UnparsedPanel[];
			graphics?: UnparsedGraphic[];
			assetCategories?: UnparsedAssetCategory[];
			mount?: UnparsedMount[];
			soundCues?: UnparsedSoundCue[];
			bundleDependencies?: UnparsedBundleDependencies;
		}
	}
// ...
	export type Manifest = Omit<PackageJSON, "mooncg"> &
		Manifest.UnparsedManifest & { transformBareModuleSpecifiers: boolean };

	/**
	 * A collection of types that comprise a `bundle`.
	 */
	export namespace Bundle {
		export type GitData =
			| undefined
			| {
					branch: string;
					hash: string;
					shortHash: string;
			  }
			| {
					branch: string;
					hash: string;
					shortHash: string;
					date: string;
					message: string;
			  };

		export type Graphic = {
			url: string;
		} & Required<Manifest.UnparsedGraphic>;

		export type Panel = Manifest.UnparsedPanel & {
			path: string;
			headerColor: string;
			bundleName: string;
			html: string;
		} & (
				| {
						dialog: false;
						workspace: string;
				  }
				| {
						dialog: true;
				  }
			) &
			(
				| {
						fullbleed: false;
						width: number;
				  }
				| {
						fullbleed: true;
				  }
			);

		export type Mount = Manifest.UnparsedMount;

		export type SoundCue = Manifest.UnparsedSoundCue;

		export type AssetCategory = Manifest.UnparsedAssetCategory;

		export type BundleDependencies = Manifest.UnparsedBundleDependencies;

		export type UnknownConfig = Record<string, unknown>;
	}

	export interface MooncgBundleConfig {
		databaseAdapter?: DatabaseAdapter;
	}

	/**
	 * The actual type of a `bundle`.
	 */
	export interface Bundle {
		name: string;
		version: string;
		license?: string;
		description?: string;
		homepage?: string;
		author?: Person;
		contributors?: Person[];
		dir: string;
		git: Bundle.GitData;
		transformBareModuleSpecifiers: boolean;
		hasAssignableSoundCues: boolean;
		hasExtension: boolean;
		config: Bundle.UnknownConfig;
		dashboard: {
			dir: string;
			panels: Bundle.Panel[];
		};
		graphics: Bundle.Graphic[];
		assetCategories: Bundle.AssetCategory[];
		mount: Bundle.Mount[];
		soundCues: Bundle.SoundCue[];
		compatibleRange?: string;
		bundleDependencies?: Bundle.BundleDependencies;
		mooncgBundleConfig: MooncgBundleConfig;
	}
// ...
	export namespace Replicant {
		export type Options<D> = OptionsNoDefault | OptionsWithDefault<D>;

		export interface OptionsNoDefault {
			persistent?: boolean;
			persistenceInterval?: number;
			schemaPath?: string;
		}

		export interface OptionsWithDefault<D> {
			persistent?: boolean;
			persistenceInterval?: number;
			schemaPath?: string;
			defaultValue: D;
		}

		export type Operation<T> = {
			path: string;
		} & ( // Objects and arrays
			| { method: "overwrite"; args: { newValue: T | undefined } }
			| { method: "delete"; args: { prop: string | number | symbol } }
			| { method: "add"; args: { prop: string; newValue: any } }
			| { method: "update"; args: { prop: string; newValue: any } }

			// Array mutator methods
			// This whole thing is gross and needs to be removed in v3
			// It is rife with unsupported cases, very easy to make bugs here
			| {
					method: "copyWithin";
					args: { mutatorArgs: Parameters<any[]["copyWithin"]> };
			  }
			| { method: "fill"; args: { mutatorArgs: Parameters<any[]["fill"]> } }
			| { method: "pop" }
			| { method: "push"; args: { mutatorArgs: Parameters<any[]["push"]> } }
			| { method: "reverse" }
			| { method: "shift" }
			| { method: "sort"; args: { mutatorArgs: Parameters<any[]["sort"]> } }
			| { method: "splice"; args: { mutatorArgs: Parameters<any[]["splice"]> } }
			| {
					method: "unshift";
					args: { mutatorArgs: Parameters<any[]["unshift"]> };
			  }
		);
	}

	/**
	 * Express middleware that can be mounted to MoonCG's webserver via `mooncg.mount`.
	 */
	export type Middleware =
		ExpressCore.ApplicationRequestHandler<express.Application>;
// ...
	export interface HandledAcknowledgement {
		handled: true;
	}

	/**
	 * An unhandled `listenFor` acknowledgement.
	 * It can safely be called/invoked.
	 */
	export interface UnhandledAcknowledgement {
		handled: false;
		(err?: any, response?: unknown): void;
	}

	/**
	 * A `listenFor` acknowledgement, sent from client to server.
	 */
	export type Acknowledgement =
		HandledAcknowledgement | UnhandledAcknowledgement;
// ...
}